K562 Gasperini reference data for CRISPR power analysis

Pre-computed pilot data from K562 chronic myelogenous leukemia cells generated using 10x Chromium technology. Contains baseline gene expression parameters and library size information for power analysis of CRISPR-based perturbation experiments.

Usage

K562_Gasperini

Format

A list with 3 elements:

baseline_expression_stats

Data frame with gene expression data (19,942 genes × 3 columns):

response_id: Character vector of Ensembl gene IDs

relative_expression: Numeric vector of relative expression levels (TPM/1e6 scale)

expression_size: Numeric vector of dispersion parameters (theta)

library_parameters

List containing:

UMI_per_cell: Maximum UMI per cell parameter (59,163)

variation: Variation parameter for PCR bias (0.397)

mapping_efficiency

Numeric. Mapping efficiency value (0.711)

Source

Paper: A Genome-wide Framework for Mapping Gene Regulation via Cellular Genetic Screens

Author and Year: Gasperini et al., 2019

Journal: Cell

Accession: GSE120861

PMID: 30612741

Details

This dataset was generated from K562 chronic myelogenous leukemia cells using single-cell RNA sequencing with 10x Chromium technology.

Cells Used in Relative Expression Estimate: All cells in high-moi condition

See also

Examples

data(K562_Gasperini)
str(K562_Gasperini)
#> List of 3
#>  $ baseline_expression_stats:'data.frame':	19942 obs. of  3 variables:
#>   ..$ response_id        : chr [1:19942] "ENSG00000243485" "ENSG00000238009" "ENSG00000241860" "ENSG00000286448" ...
#>   ..$ relative_expression: num [1:19942] 1.50e-07 1.55e-06 7.33e-07 9.31e-07 7.43e-06 ...
#>   ..$ expression_size    : num [1:19942] 0.3556 2.4156 7.7329 0.0341 3.4513 ...
#>  $ library_parameters       :List of 2
#>   ..$ UMI_per_cell: num 59163
#>   ..$ variation   : num 0.397
#>  $ mapping_efficiency       : num 0.711