Compute Monte Carlo test statistics for power analysis with random effect sizes

Compute Monte Carlo test statistics for power analysis with random effect sizes

Usage

compute_monte_carlo_teststat_cpp(
  fc_expression_df,
  library_size,
  num_trt_cells,
  num_cntrl_cells
)

Arguments

fc_expression_df

Data frame with fold change and expression information

library_size

Library size parameter

num_trt_cells

Number of treatment cells

num_cntrl_cells

Number of control cells

Value

List with Monte Carlo mean and standard deviation vectors