Input checking function for compute_power_posthoc

Input checking function for compute_power_posthoc

Usage

input_check_posthoc(
  discovery_pairs = NULL,
  cells_per_grna = NULL,
  baseline_expression_stats = NULL,
  control_group = NULL,
  fold_change_mean = NULL,
  fold_change_sd = NULL,
  num_total_cells = NULL,
  cutoff = NULL,
  n_nonzero_trt_thresh = NULL,
  n_nonzero_cntrl_thresh = NULL,
  side = NULL,
  multiple_testing_method = NULL,
  multiple_testing_alpha = NULL
)

Arguments

discovery_pairs

A data frame specifying which element-gene pairs to consider, with columns grna_target and response_id

cells_per_grna

A data frame specifying how many cells contain each gRNA, with columns grna_id, grna_target, and num_cells

baseline_expression_stats

A data frame specifying the baseline expression statistics for each gene, with columns response_id, expression_mean, and expression_size

control_group

A character string specifying the control group, either "complement" or "nt_cells"

fold_change_mean

A numeric value to use for mean effect size for all element-gene pairs

fold_change_sd

A numeric value to use for standard deviation of effect size for all element-gene pairs

num_total_cells

(Required only if control_group == "complement") A positive integer specifying the total number of cells in the experiment

cutoff

(Optional) A numeric value between 0 and 1 to use as the p-value cutoff

n_nonzero_trt_thresh

(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7

n_nonzero_cntrl_thresh

(Optional) An integer specifying the sceptre QC parameter of the same name; defaults to 7

side

(Optional) A character string specifying the side of the test, either "left", "right", or "both"; defaults to "both"

multiple_testing_method

(Optional) A character string specifying the multiple testing correction method to use, either "BH" or "bonferroni"; defaults to "BH"

multiple_testing_alpha

(Optional) A numeric value between 0 and 1 specifying the alpha level for multiple testing correction; defaults to 0.1